requirements.txt

fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
requests
